Forge update 0.1: rebalance tool progression vertical - Added forging furnace - Added moulds for tool heads - Added forging - Removed old tool crafting recipes if forge mod installed - Added arbitrary world draw to modding abi functions - Extended GUI for modding support + animated slots - Added gold tools - Added copper tools - Added pottery table - Cleaned up some code - Improved performance

+++ mod-api/src/ids.rs
@@ -2,7 +2,7 @@
 
 use serde::{Deserialize, Serialize};
 
-/// A runtime block id — raw `u8` into the engine's registry. Dynamic content is
+/// A runtime block id — raw `u16` into the engine's registry. Dynamic content is
 /// NAME-addressed (`mod_id:name` keys in the pack catalogs assign ids at load),
 /// so numeric ids are stable within a session but never across sessions or
 /// saves; mods must not persist them. Resolve ids from names at `mod_init` time
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 ///
 /// [`HostCall::ResolveBlock`]: crate::HostCall::ResolveBlock
 #[derive(Serialize, Deserialize, Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct BlockId(pub u8);
+pub struct BlockId(pub u16);
 
 impl BlockId {
     /// Air is engine id 0 — the one numeric id frozen by contract (worldgen
@@ -20,7 +20,7 @@ impl BlockId {
 
 /// A runtime item id — same contract as [`BlockId`].
 #[derive(Serialize, Deserialize, Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct ItemId(pub u8);
+pub struct ItemId(pub u16);
 
 /// A runtime mob SPECIES id — same contract as [`BlockId`]. This identifies a
 /// kind (`"petramond:sheep"`), never a live instance; live mobs are addressed
